fix: Update SchoolRepository to use Long for ID type instead of Integer

+++ src/main/java/rw/gov/mineduc/qamis/integration/repository/SchoolRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import rw.gov.mineduc.qamis.integration.model.School;
 
-public interface SchoolRepository extends JpaRepository<School, Integer>, JpaSpecificationExecutor<School> {
+public interface SchoolRepository extends JpaRepository<School, Long>, JpaSpecificationExecutor<School> {
 }
